fix: Timetable Context를 SelectSchedule에서 주입하는 것이 아니라 Timetable 컴포넌트에서 주입함
