+++ src/pages/selectSchedule/components/SelectScheduleTable.tsx
@@ -1,6 +1,3 @@
-import { useState } from 'react';
-
-import { SelectedSlotType, TimetableContext } from 'components/timetableComponents/context';
 import Timetable from 'components/timetableComponents/Timetable';
 import { ColumnStructure, TimetableStructure } from 'components/timetableComponents/types';
 
@@ -14,8 +11,6 @@ import { StepBtnsType, StepSlotsType } from '../types';
 
 function SelectScheduleTable({ timeSlots, availableDates }: TimetableStructure) {
   const { scheduleStep } = useScheduleStepContext();
-  const [startSlot, setStartSlot] = useState<string | undefined>(undefined);
-  const [selectedSlots, setSelectedSlots] = useState<SelectedSlotType>({});
 
   const stepSlots: StepSlotsType = {
     selectTimeSlot: ({ date, timeSlots }: ColumnStructure) => (
@@ -27,28 +22,20 @@ function SelectScheduleTable({ timeSlots, availableDates }: TimetableStructure) 
   };
   const stepSlot = stepSlots[scheduleStep];
 
-  const isValidSelection = Object.keys(selectedSlots).length !== 0;
   const stepBtns: StepBtnsType = {
-    selectTimeSlot: <TimeSlotCta isValidSelection={isValidSelection} />,
+    selectTimeSlot: <TimeSlotCta />,
     selectPriority: <PriorityCta />,
   };
   const stepBtn = stepBtns[scheduleStep];
 
   return (
-    <TimetableContext.Provider
-      value={{
-        startSlot,
-        setStartSlot,
-        selectedSlots,
-        setSelectedSlots,
-      }}
-    >
+    <>
       <Timetable timeSlots={timeSlots} availableDates={availableDates}>
         {stepSlot}
       </Timetable>
       {scheduleStep === 'selectPriority' && <PriorityDropdown />}
       {stepBtn}
-    </TimetableContext.Provider>
+    </>
   );
 }
